requirements.txt
# Core dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-multipart==0.0.12

# Image processing
pillow==11.2.1
numpy==2.2.0

# Scheduling
apscheduler==3.10.4
pytz==2025.2

# HTTP client
requests==2.32.3
httpx==0.27.2

# Configuration
pyyaml==6.0.2

# Utilities
python-dateutil==2.9.0
psutil==6.1.0
